requirements.txt
# backend/requirements.txt
fastapi
uvicorn
supabase
google-generativeai  # Replaced openai
pydantic
python-dotenv